incomplete assignment again

+++ app/controllers/comments_controller.rb
@@ -4,26 +4,50 @@ class CommentsController < ApplicationController
 
 
   def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
 
-    if comment.save
-      flash[:notice] = "Comment saved successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @post]
+    if params[:post_id]
+      @parent = Post.find(params[:post_id])
+      comment = @parent.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
+    elsif params[:topic_id]
+      @parant = Topic.find(params[:topic_id])
+      comment = @parent.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
+    if params[:post_id]
+      @parent = Post.find(params[:post_id])
+      comment = @parent.comments.find(params[:id])
 
-    if comment.destroy
-      flash[:notice] = "Comment was deleted"
-      redirect_to [@post.topic, @post]
+      if comment.destroy
+        flash[:notice] = "Comment was deleted"
+        redirect_to [@post.topic, @post]
+      end
+    elsif params[:topic_id]
+      @parent = Post.find(params[:post_id])
+      comment = @parent.comments.find(params[:id])
+      if comment.destroy
+        flash[:notice] = "Comment was deleted"
+        redirect_to [@post.topic, @post]
+      end
     end
   end
